+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### 2024-02-08
 
+### Added
+
+- You can now link to specific items, i.e. https://remnant2toolkit.com/item/bandofaccord. Share links have been added to item cards and item info.
+
 ### Changed
 
 - Updated the Item Lookup page filters to be used in the URL, allowing you to share a link to a specific set of tags with the filters applied.